# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Watchdog Fleet & Telemetry REST API
  description: Versioned REST API for local node observability, centralized fleet management, telemetry ingestion, and cluster health monitoring.
  version: 1.0.0
servers:
  - url: https://fleet.internal:8443
    description: Production Fleet Server
  - url: http://127.0.0.1:8443
    description: Local Agent Server
paths:
...
  /api/v1/fleet:
    get:
      summary: List Fleet Nodes
      description: Queries registered fleet nodes with filtering, pagination, and sorting.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [healthy, warning, critical, stale, offline, unknown]
          description: Filter nodes by status
        - in: query
          name: search
          schema:
            type: string
          description: Search by node ID, hostname, or IP address
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: sort_by
          schema:
            type: string
            enum: [hostname, last_heartbeat, cpu, memory, status]
        - in: query
          name: sort_direction
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: since
          schema:
            type: string
          description: Active since RFC3339 timestamp or duration (e.g. 10m, 2h)
      responses:
        '200':
          description: Filtered list of fleet nodes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FleetListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/fleet/{node_id}:
    get:
      summary: Get Node Detail
      description: Retrieves comprehensive status, hardware specs, diagnostics, and recent telemetry for a single node.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
          description: Unique node identifier
      responses:
        '200':
          description: Detailed node information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Deregister Node
      description: Removes a node from the fleet registry.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Node successfully removed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
                  node_id:
                    type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
  responses:
...
    UnauthorizedError:
      description: Missing or invalid authentication token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIErrorResponse'
    NotFoundError:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIErrorResponse'
...
  schemas:
    APIErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: INVALID_TOKEN
            message:
              type: string
              example: Bearer authentication token is missing or invalid
            request_id:
              type: string
              example: c6a2b8e4-18c7-43cf-8ef3-d6c62c2f42a1
            timestamp:
              type: string
              format: date-time
    NodeIdentity:
      type: object
      required:
        - node_id
        - hostname
        - os
        - arch
      properties:
        node_id:
          type: string
          example: c56a4180-65aa-42ec-a945-5fd21dec0538
        hostname:
          type: string
          example: prod-worker-01.internal
        os:
          type: string
          example: linux
        platform:
          type: string
          example: ubuntu
        platform_version:
          type: string
          example: "22.04"
        kernel_version:
          type: string
          example: "5.15.0-101-generic"
        arch:
          type: string
          example: amd64
        cpu_cores:
          type: integer
          example: 16
        total_memory:
          type: integer
          format: int64
          example: 34359738368
        ip_addresses:
          type: array
          items:
            type: string
        mac_addresses:
          type: array
          items:
            type: string
        version:
          type: string
          example: 1.0.0
        tags:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
    TelemetrySubmission:
      type: object
      required:
        - node_id
        - timestamp
      properties:
        node_id:
          type: string
        timestamp:
          type: string
          format: date-time
        snapshot:
          type: object
        diagnostic_status:
          type: string
          example: HEALTHY
        active_alerts_count:
          type: integer
          example: 0
...
    FleetListResponse:
      type: object
      properties:
        total:
          type: integer
        count:
          type: integer
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/FleetNode'
    NodeDetailResponse:
      type: object
      properties:
        node:
          $ref: '#/components/schemas/FleetNode'
        recent_telemetry:
          type: array
          items:
            $ref: '#/components/schemas/TelemetrySubmission'
        active_alerts:
          type: array
          items:
            type: object
    FleetNode:
      type: object
      properties:
        identity:
          $ref: '#/components/schemas/NodeIdentity'
        status:
          type: string
          enum: [healthy, warning, critical, stale, offline, unknown]
        status_message:
          type: string
        registered_at:
          type: string
          format: date-time
        last_heartbeat:
          type: string
          format: date-time
        last_telemetry:
          type: string
          format: date-time
        summary:
          type: object
        metadata:
          type: object
          additionalProperties:
            type: string